+++ DestinyArmourSelector/ArmourPieceCreator.cs
@@ -2,6 +2,7 @@
 
 namespace DestinyArmourSelector
 {
+    using Interfaces;
     using System.Collections.Generic;
     using System.IO;
     using System.Threading.Tasks;
