+++ DestinyArmourSelector/Interfaces/ISynergyCalculator.cs
@@ -1,0 +1,9 @@
+﻿// Copyright Small & Fast 2018
+
+namespace DestinyArmourSelector.Interfaces
+{
+    public interface ISynergyCalculator
+    {
+        string CalculateSynergy(PerkGroup primaryPerks, PerkGroup secondaryPerks);
+    }
+}
