+++ DestinyArmourSelector/Program.cs
@@ -2,6 +2,7 @@
 
 namespace DestinyArmourSelector
 {
+    using Interfaces;
     using System;
     using System.Collections.Generic;
     using System.Linq;
